Rename the files to be only linux
